feat(cluster): persistent cluster sub-nav rail; share the rail component
Extract the responsive sub-nav (left rail on desktop, dropdown when small, with
expandable sub-items) into a shared SubNavRail so the org and cluster layouts
use the same plumbing. Add ClusterPageLayout with an Overview / Instances /
Scaling / Domains rail (permission-gated; domains hides for self-managed) and
apply it across those cluster-scoped pages so the rail persists — the in-cluster
studio (apps, databases, …) keeps its own top nav.

The cluster home now renders inside the rail (its old horizontal "Manage" tiles
are gone — the rail replaces them); Instances, Scaling, and Domains adopt it too
(StartingUp stays plain — nothing to manage pre-provision).
