+++ src/features/cluster/domains/Page.tsx
@@ -1,10 +1,10 @@
-import { NestedContentWithSubNavMenu } from '@/components/NestedContentWithSubNavMenu';
+import { ClusterContentWithSubNavMenu } from '@/features/cluster/components/ClusterContentWithSubNavMenu';
 import { DomainsManagement } from './Management';
 
 export function DomainsPage() {
 	return (
-		<NestedContentWithSubNavMenu className="flex flex-col justify-start max-w-4xl">
+		<ClusterContentWithSubNavMenu className="flex flex-col justify-start max-w-4xl">
 			<DomainsManagement />
-		</NestedContentWithSubNavMenu>
+		</ClusterContentWithSubNavMenu>
 	);
 }
